requirements.txt
facenet_pytorch
torch
torchvision
pandas
opencv-python
pillow
